backend/requirements.txt
# Smart Civic Complaint Analyzer - Backend Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.29
psycopg2-binary==2.9.9       # PostgreSQL driver
alembic==1.13.1               # Database migrations (optional but good practice)

# Data Validation
pydantic==2.6.4

# Environment Variables
python-dotenv==1.0.1

# Testing
httpx==0.27.0                 # For testing FastAPI
pytest==8.1.1

# ── Optional: For Advanced AI Mode ───────────────────────────────────────────
# Uncomment these if you want to use the Hugging Face model (requires ~2GB disk)
# transformers==4.39.3
# torch==2.2.2
# numpy==1.26.4